+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Roadmap
 
-- [ ] Log to output window
+- [x] Log to output window
 
 Features that have a checkmark are complete and available for
 download in the